Fix type conversion warnings
Re #23027

+++ qt/scientific_interfaces/ISISReflectometry/Reduction/Group.h
@@ -88,8 +88,8 @@ void mergeRowsInto(Group &intoHere, Group const &fromHere, int groupIndex,
   }
 }
 
-size_t totalItems(Group const &group);
-size_t completedItems(Group const &group);
+int totalItems(Group const &group);
+int completedItems(Group const &group);
 
 } // namespace CustomInterfaces
 } // namespace MantidQt
